+++ test/test.js
@@ -94,12 +94,3 @@ test('ids spanning 1ms boundary are 100ns apart', function() {
   var dt = parseInt(after, 16) - parseInt(before, 16);
   assert(dt === 1, 'Ids spanning 1ms boundary are 100ns apart');
 });
-
-test('parse/unparse', function() {
-  var id = '00112233445566778899aabbccddeeff';
-  assert(uuid.unparse(uuid.parse(id.substr(0,10))) ==
-    '00112233-4400-0000-0000-000000000000', 'Short parse');
-  assert(uuid.unparse(uuid.parse('(this is the uuid -> ' + id + id)) ==
-    '00112233-4455-6677-8899-aabbccddeeff', 'Dirty parse');
-});
-